% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfl_colour.R
\name{rfl_colour}
\alias{rfl_colour}
\title{Return Colour from the RFL style (in Hexidecimal Format)}
\usage{
rfl_colour(colour, alpha = 1)
}
\arguments{
\item{colour}{A colour in the RFL pallete (type rfl_colors() for lists).}
}
\value{
A string in Hexidecimal format for use in graphics & themes
}
\description{
Return Colour from the RFL style (in Hexidecimal Format)
}
\examples{
ggplot(data, aes(x=xaxis, y=yaxis)) +
geom_line(colour=rfl_colour("blue"))
rfl_colour("green")
}
