% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfl_colour.R
\name{rfl_colour_pal}
\alias{rfl_colour_pal}
\title{Internal comand to deal with the RFL colour pallete}
\usage{
rfl_colour_pal(palette = "main", reverse = FALSE, ...)
}
\description{
Internal comand to deal with the RFL colour pallete
}
