% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfl_colour.R
\name{rfl_colours}
\alias{rfl_colours}
\title{Return a List of Colours in the RFL style (in Hexidecimal Format)}
\usage{
rfl_colours(details = FALSE)
}
\arguments{
\item{details}{A logical defaulting to FALSE. Provides colour names and RGB values}
}
\value{
A vector of colour names to use in rfl_colour() or a dataframe with details.
}
\description{
Return a List of Colours in the RFL style (in Hexidecimal Format)
}
\examples{
rfl_colours()
rfl_colours(details=TRUE)
}
