% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfl_colour.R
\docType{data}
\name{rfl_pallete}
\alias{rfl_pallete}
\title{show different available colour palletes}
\format{An object of class \code{list} of length 4.}
\usage{
rfl_pallete
}
\description{
show different available colour palletes
}
\keyword{datasets}
