% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfl_colour.R
\name{rfl_scale_colour}
\alias{rfl_scale_colour}
\title{use a colour scale from the Royal Free colour pallete for 'colour' attributes}
\usage{
rfl_scale_colour(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{discrete}{A logical to set a discrete or coninuious colour scale}

\item{reverse}{A logical to reverse the colour scale}

\item{pallete}{A string denoting the pallete to use, options "main", "bright" and "full"}
}
\value{
ggplot2::discrete_scale or scale_color_gradient object
}
\description{
use a colour scale from the Royal Free colour pallete for 'colour' attributes
}
\examples{
geom_line() +
scale_color_rfl("main", reverse=TRUE)
}
