% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfl_style.R
\name{rfl_show_axis_titles}
\alias{rfl_show_axis_titles}
\title{Show Axis Titles on an rfl_style() Styled ggplot2 graphic}
\usage{
rfl_show_axis_titles()
}
\value{
ggplot::theme object to be added (+) to ggplot2 graphic
}
\description{
Show Axis Titles on an rfl_style() Styled ggplot2 graphic
}
\examples{
rfl_show_axis_titles()

ggplot(data, aes(x=xaxis, y=yaxis)) +
  geom_line() +
  rfl_style() +
  labs(x="something", y="something else") +
  rfl_show_axis_titles()
}
