% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfl_style.R
\name{rfl_style}
\alias{rfl_style}
\title{Add RFL styling to ggplot graphics}
\usage{
rfl_style()
}
\value{
ggplot::theme object to be added (+) to ggplot2 graphic
}
\description{
Add RFL styling to ggplot graphics
}
\examples{
rfl_style()

ggplot(data, aes(x=xaxis, y=yaxis)) +
geom_line() +
rfl_style()
}
